Add back controllers
Added back a base controller class, but the utilities are no longer
bound to triggers. Instead, if you want to create a trigger from them,
it has to be done manually. This was done in the spirit of the new way
that triggers work in WPILib, breaking them away from controller inputs